feat(img): support des types étendus Garmin (≥ 0x100) — forêts, bâtiments, POIs
Implémente les 3 sections RGN étendues (areas, lines, points) avec leur
format binaire spécifique, les sections TRE associées (extTypeOffsets,
extTypeOverviews), et adapte le parser/writer pour les types sur 32 bits.

La carte passe de ~4% des polygones visibles à 100%, récupérant forêts,
bâtiments, occupation du sol et tous les POIs. Validé visuellement dans
QMapShack avec TYP I2023100.typ.

+++ imgforge-cli/src/img/map_object.rs
@@ -4,7 +4,7 @@ use super::coord::Coord;
 
 /// Common properties for all map objects (points, polylines, polygons)
 pub trait MapObject {
-    fn type_code(&self) -> u16;
+    fn type_code(&self) -> u32;
     fn sub_type(&self) -> u8 { 0 }
     fn label_offset(&self) -> u32;
     fn coords(&self) -> &[Coord];
